Refactor: 전반적인 코드 리팩토링 #13

+++ src/main/java/com/external/auth/util/JwtUtil.java
@@ -5,8 +5,6 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-
-import java.math.BigInteger;
 import java.util.Date;
 
 @Component
@@ -18,7 +16,7 @@ public class JwtUtil {
     @Value("${jwt.expiration_ms}")
     private long expirationMs;
 
-    public String generateToken(BigInteger userId) {
+    public String generateToken(Long userId) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + expirationMs);
 
@@ -30,13 +28,13 @@ public class JwtUtil {
                 .compact();
     }
 
-    public BigInteger getUserId(String token) {
+    public Long getUserId(String token) {
         Claims claims = Jwts.parser()
                 .setSigningKey(secretKey)
                 .parseClaimsJws(token)
                 .getBody();
 
-        return BigInteger.valueOf(Long.parseLong(claims.getSubject()));
+        return Long.parseLong(claims.getSubject());
     }
 
     public boolean validateToken(String token) {
